+++ ownCloud/Settings/AutoUploadSettingsSection.swift
@@ -167,9 +167,7 @@ class AutoUploadSettingsSection: SettingsSection {
 			instantUploadPhotosRow = StaticTableViewRow(switchWithAction: { [weak self] (_, sender) in
 				if let convertSwitch = sender as? UISwitch {
 					self?.changeAndRequestPhotoLibraryAccessForOption(optionSwitch: convertSwitch, completion: { (switchState) in
-						OnMainThread {
-							self?.setupPhotoAutoUpload(enabled: switchState)
-						}
+						self?.setupPhotoAutoUpload(enabled: switchState)
 					})
 				}
 				}, title: "Auto Upload Photos".localized, value: self.userDefaults.instantUploadPhotos, identifier: "auto-upload-photos")
@@ -177,9 +175,7 @@ class AutoUploadSettingsSection: SettingsSection {
 			instantUploadVideosRow = StaticTableViewRow(switchWithAction: { [weak self] (_, sender) in
 				if let convertSwitch = sender as? UISwitch {
 					self?.changeAndRequestPhotoLibraryAccessForOption(optionSwitch: convertSwitch, completion: { (switchState) in
-						OnMainThread {
-							self?.setupVideoAutoUpload(enabled: switchState)
-						}
+						self?.setupVideoAutoUpload(enabled: switchState)
 					})
 				}
 				}, title: "Auto Upload Videos".localized, value: self.userDefaults.instantUploadVideos, identifier: "auto-upload-videos")
@@ -200,7 +196,6 @@ class AutoUploadSettingsSection: SettingsSection {
 	}
 
 	private func setupPhotoAutoUpload(enabled:Bool) {
-
 		if !enabled {
 			userDefaults.resetInstantPhotoUploadConfiguration()
 			postSettingsChangedNotification()
@@ -210,6 +205,8 @@ class AutoUploadSettingsSection: SettingsSection {
 			userDefaults.instantUploadPhotosAfter = Date()
 			if userDefaults.instantPhotoUploadPath == nil || userDefaults.instantPhotoUploadBookmarkUUID == nil {
 				showAccountSelectionViewController(for: .photo)
+			} else {
+				updateDynamicUI()
 			}
 		}
 	}
@@ -224,7 +221,9 @@ class AutoUploadSettingsSection: SettingsSection {
 			userDefaults.instantUploadVideosAfter = Date()
 			if userDefaults.instantVideoUploadPath == nil || userDefaults.instantVideoUploadBookmarkUUID == nil {
 				showAccountSelectionViewController(for: .video)
-			}
+			} else {
+			   updateDynamicUI()
+		   }
 		}
 	}
 
@@ -388,7 +387,7 @@ class AutoUploadSettingsSection: SettingsSection {
 											guard let core = core, error == nil else { return }
 
 											OnMainThread {
-												let directoryPickerViewController = ClientDirectoryPickerViewController(core: core, path: "/", selectButtonTitle: "Select Upload Path".localized, avoidConflictsWith: [], choiceHandler: { (selectedDirectory) in
+												let directoryPickerViewController = ClientDirectoryPickerViewController(core: core, path: "/", selectButtonTitle: "Select Upload Path".localized, avoidConflictsWith: [], choiceHandler: { (selectedDirectory, _) in
 													OCCoreManager.shared.returnCore(for: bookmark, completionHandler: nil)
 													completion(selectedDirectory)
 												})