+++ ownCloudAppShared/Client/File Lists/ClientQueryViewController.swift
@@ -215,12 +215,6 @@ open class ClientQueryViewController: QueryFileListTableViewController, UIDropIn
 
 	}
 
-	open func tableView(_: UITableView, dragSessionDidEnd: UIDragSession) {
-		if !self.tableView.isEditing {
-			removeToolbar()
-		}
-	}
-
 	// MARK: - UIBarButtonItem Drop Delegate
 
 	open func dropInteraction(_ interaction: UIDropInteraction, canHandle session: UIDropSession) -> Bool {
@@ -384,14 +378,14 @@ open class ClientQueryViewController: QueryFileListTableViewController, UIDropIn
 	// MARK: - Reloads
 	open override func restoreSelectionAfterTableReload() {
 		// Restore previously selected items
-		if tableView.isEditing && selectedItemIds.count > 0 {
-			var selectedItems = [OCItem]()
-			for row in 0..<self.items.count {
-				if let itemLocalID = self.items[row].localID as OCLocalID? {
-					if selectedItemIds.contains(itemLocalID) {
-						selectedItems.append(self.items[row])
-						self.tableView.selectRow(at: IndexPath(row: row, section: 0), animated: false, scrollPosition: .none)
-					}
+		guard tableView.isEditing else { return }
+
+		guard selectedItemIds.count > 0 else { return }
+
+		for row in 0..<self.items.count {
+			if let itemLocalID = self.items[row].localID as OCLocalID? {
+				if selectedItemIds.contains(itemLocalID) {
+					self.tableView.selectRow(at: IndexPath(row: row, section: 0), animated: false, scrollPosition: .none)
 				}
 			}
 		}
@@ -556,6 +550,12 @@ extension ClientQueryViewController: UITableViewDragDelegate {
 		}
 
 		return []
+	}
+
+	public func tableView(_: UITableView, dragSessionDidEnd: UIDragSession) {
+		if !self.tableView.isEditing {
+			removeToolbar()
+		}
 	}
 
 	public func itemForDragging(draggingValue : OCItemDraggingValue) -> UIDragItem? {